# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Elma API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/internal/conversations/{cid}": {
      "get": {
        "tags": [
          "Worker"
        ],
        "summary": "Config",
        "operationId": "config_api_internal_conversations__cid__get",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/internal/conversations/{cid}/messages": {
      "post": {
        "tags": [
          "Worker"
        ],
        "summary": "Message",
        "operationId": "message_api_internal_conversations__cid__messages_post",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MessageEvent"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/internal/conversations/{cid}/search": {
      "post": {
        "tags": [
          "Worker"
        ],
        "summary": "Search",
        "operationId": "search_api_internal_conversations__cid__search_post",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SearchInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/internal/conversations/{cid}/close": {
      "post": {
        "tags": [
          "Worker"
        ],
        "summary": "Close",
        "operationId": "close_api_internal_conversations__cid__close_post",
        "parameters": [
          {
            "name": "cid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Cid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CloseInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/organizations/{org_id}/usage": {
      "get": {
        "tags": [
          "Usage"
        ],
        "summary": "Usage",
        "operationId": "usage_api_organizations__org_id__usage_get",
        "parameters": [
          {
            "name": "org_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Org Id"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Usage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CloseInput": {
        "properties": {
          "run_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "completed",
              "error"
            ],
            "title": "Status",
            "default": "completed"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "CloseInput"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MessageEvent": {
        "properties": {
          "id": {
            "type": "string",
            "maxLength": 150,
            "title": "Id"
          },
          "role": {
            "type": "string",
            "enum": [
              "user",
              "assistant"
            ],
            "title": "Role"
          },
          "content": {
            "type": "string",
            "maxLength": 60000,
            "title": "Content"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "role",
          "content"
        ],
        "title": "MessageEvent"
      },
...
      "SearchInput": {
        "properties": {
          "query": {
            "type": "string",
            "maxLength": 12000,
            "minLength": 1,
            "title": "Query"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "query"
        ],
        "title": "SearchInput"
      },
...
      "Usage": {
        "properties": {
          "totals": {
            "$ref": "#/components/schemas/UsageTotals"
          },
          "entries": {
            "items": {
              "$ref": "#/components/schemas/UsageEntry"
            },
            "type": "array",
            "title": "Entries"
          },
          "currency": {
            "type": "string",
            "title": "Currency"
          },
          "pricing": {
            "type": "string",
            "title": "Pricing"
          }
        },
        "type": "object",
        "required": [
          "totals",
          "entries",
          "currency",
          "pricing"
        ],
        "title": "Usage"
      },
      "UsageEntry": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "agent_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Id"
          },
          "conversation_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Conversation Id"
          },
          "agent_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Name"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "operation": {
            "type": "string",
            "title": "Operation"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "quantities": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Quantities"
          },
          "tariff": {
            "additionalProperties": true,
            "type": "object",
            "title": "Tariff"
          },
          "cost": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cost"
          },
          "price": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Price"
          },
          "raw_usage": {
            "additionalProperties": true,
            "type": "object",
            "title": "Raw Usage"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "agent_id",
          "conversation_id",
          "agent_name",
          "provider",
          "request_id",
          "operation",
          "model",
          "quantities",
          "tariff",
          "cost",
          "price",
          "raw_usage",
          "created_at"
        ],
        "title": "UsageEntry"
      },
      "UsageTotals": {
        "properties": {
          "cost": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Cost"
          },
          "price": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Price"
          },
          "calls": {
            "type": "integer",
            "title": "Calls"
          },
          "unpriced": {
            "type": "integer",
            "title": "Unpriced"
          }
        },
        "type": "object",
        "required": [
          "cost",
          "price",
          "calls",
          "unpriced"
        ],
        "title": "UsageTotals"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}